Move health radar to stress panel
